change Handleseed arge from int32 to uint64 and parse as hex

+++ include/dxf/file_element.hpp
@@ -9,9 +9,9 @@ namespace dxf{
    struct file_image_t;
    struct entities_t;
    struct tables_t;
-      
+
    struct file_element_t{
-      int32_t get_handle() const { return m_handle;}
+      uint64_t get_handle() const { return m_handle;}
       void set_file_image(file_image_t &);
       file_image_t & get_file_image() { assert(m_file_image != nullptr); return *m_file_image;}
       file_image_t const & get_file_image()const { assert(m_file_image != nullptr);return *m_file_image;}
@@ -28,7 +28,7 @@ namespace dxf{
       friend struct entities_t;
       friend struct tables_t;
       file_image_t * m_file_image;
-      int32_t m_handle;
+      uint64_t m_handle;
    };
 
 }
